lesson15/conversational-agent/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
google-generativeai==0.8.3
python-dotenv==1.0.1
aiosqlite==0.20.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
